restore working addon search

+++ include/admin/admin_addon_install.php
@@ -258,14 +258,14 @@ class admin_addon_install extends admin_addons_tool{
 		}
 
 		//invalid response?
-		//if( strpos($result,'a:') !== 0 ){
+		if( strpos($result,'a:') !== 0 ){
 			if( $use_cache ) unlink($cache_file);
 			$debug				= array();
 			$debug['Two']		= substr($result,0,2);
 			$debug['Twotr']		= substr(trim($result),0,2);
 			echo '<p>'.gpRemoteGet::Debug('Sorry, data not fetched',$debug).'</p>';
 			return;
-		//}
+		}
 
 		$data = @unserialize($result);
